Improved toast messages. Improved signup process.
Now username is also checked for sign up.
Also, phone is no longer sent to the database as when it is set as the registeration email, it blocks the usage of such email address even after the formet user changes it. (Phone data can't be changed and database won't allow duplicate phone numbers either.)